client: Fixes for x86 platform

+++ include/elliptics/interface.h
@@ -549,7 +549,7 @@ static inline char *dnet_print_time(const struct dnet_time *t)
 	localtime_r((time_t *)&t->tsec, &tm);
 	strftime(str, sizeof(str), "%F %R:%S", &tm);
 
-	snprintf(__dnet_print_time, sizeof(__dnet_print_time), "%s.%06lu", str, t->tnsec / 1000);
+	snprintf(__dnet_print_time, sizeof(__dnet_print_time), "%s.%06llu", str, (long long unsigned) t->tnsec / 1000);
 	return __dnet_print_time;
 }
